92-Reverse-Linked-List-II/92-Reverse-Linked-List-II_test.go
package leetcode

import (
	"reflect"
	"testing"
)

func Test_reverseBetween(t *testing.T) {
	type args struct {
		head  *ListNode
		left  int
		right int
	}
	tests := []struct {
		name string
		args args
		want *ListNode
	}{
		{
			name: "example 1",
			args: args{
				head:  &ListNode{1, &ListNode{2, &ListNode{3, &ListNode{4, &ListNode{5, nil}}}}},
				left:  2,
				right: 4,
			},
			want: &ListNode{1, &ListNode{4, &ListNode{3, &ListNode{2, &ListNode{5, nil}}}}},
		},
		{
			name: "example 2",
			args: args{
				head:  &ListNode{5, nil},
				left:  1,
				right: 1,
			},
			want: &ListNode{5, nil},
		},
		{
			name: "full reverse",
			args: args{
				head:  &ListNode{1, &ListNode{2, &ListNode{3, &ListNode{4, &ListNode{5, nil}}}}},
				left:  1,
				right: 5,
			},
			want: &ListNode{5, &ListNode{4, &ListNode{3, &ListNode{2, &ListNode{1, nil}}}}},
		},
		{
			name: "partial reverse at start",
			args: args{
				head:  &ListNode{1, &ListNode{2, &ListNode{3, &ListNode{4, &ListNode{5, nil}}}}},
				left:  1,
				right: 3,
			},
			want: &ListNode{3, &ListNode{2, &ListNode{1, &ListNode{4, &ListNode{5, nil}}}}},
		},
		{
			name: "partial reverse at end",
			args: args{
				head:  &ListNode{1, &ListNode{2, &ListNode{3, &ListNode{4, &ListNode{5, nil}}}}},
				left:  3,
				right: 5,
			},
			want: &ListNode{1, &ListNode{2, &ListNode{5, &ListNode{4, &ListNode{3, nil}}}}},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := reverseBetween(tt.args.head, tt.args.left, tt.args.right); !reflect.DeepEqual(
				got,
				tt.want,
			) {
				t.Errorf("reverseBetween() = %v, want %v", got, tt.want)
			}
		})
	}
}
